requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
channels==4.2.0
daphne==4.1.2
pandas==2.2.3
numpy==2.1.3
Pillow==11.0.0
pyttsx3==2.98
python-dotenv==1.0.1

# LLM runtime (Phase 7)
torch>=2.6.0
transformers>=4.48.0
accelerate>=1.2.1
